moved around the the data to the correct files
